chore(suite): create notifications folder
