+++ packages/suite/src/components/suite/SwitchDevice/components/DeviceItem/components/DeviceHeaderButton/index.tsx
@@ -3,8 +3,7 @@ import styled from 'styled-components';
 import * as deviceUtils from '@suite-common/suite-utils';
 import { useTheme, Icon } from '@trezor/components';
 
-import { Translation } from 'src/components/suite';
-import NotificationCard from 'src/components/suite/NotificationCard'; // on purpose to avoid hacky sc overriding
+import { NotificationCard, Translation } from 'src/components/suite';
 import { TrezorDevice } from 'src/types/suite';
 
 const GrayNotificationCard = styled(NotificationCard)`
