+++ packages/suite/src/components/suite/notifications/NotificationRenderer/ConditionalActionRenderer.tsx
@@ -1,8 +1,9 @@
 import { ReactNode } from 'react';
 import styled from 'styled-components';
+
 import { variables } from '@trezor/components';
 import type { ExtendedMessageDescriptor } from 'src/types/suite';
-import type { NotificationRendererProps } from '../types';
+import type { NotificationRendererProps } from 'src/components/suite';
 
 const Header = styled.div`
     font-weight: ${variables.FONT_WEIGHT.MEDIUM};
@@ -25,7 +26,7 @@ type ConditionalActionRendererProps = NotificationRendererProps & {
     onCancel: () => void;
 };
 
-const ConditionalActionRenderer = ({
+export const ConditionalActionRenderer = ({
     header,
     body,
     actionAllowed,
@@ -58,5 +59,3 @@ const ConditionalActionRenderer = ({
         />
     );
 };
-
-export default ConditionalActionRenderer;
