+++ packages/suite/src/middlewares/suite/toastMiddleware.ts
@@ -1,8 +1,9 @@
+import { MiddlewareAPI } from 'redux';
+import { toast } from 'react-toastify';
+
 import { notificationsActions } from '@suite-common/toast-notifications';
 import { renderToast } from 'src/components/suite';
 import { Action, AppState, Dispatch } from 'src/types/suite';
-import { toast } from 'react-toastify';
-import { MiddlewareAPI } from 'redux';
 
 /*
  * Middleware for toast notifications.
