+++ packages/suite/src/utils/suite/notification.ts
@@ -1,7 +1,7 @@
 import { NotificationEntry } from '@suite-common/toast-notifications';
-import { ToastNotificationVariant, AppState } from 'src/types/suite';
 import { colors } from '@trezor/components';
-import type { NotificationViewProps } from 'src/components/suite/NotificationRenderer';
+import { ToastNotificationVariant, AppState } from 'src/types/suite';
+import type { NotificationViewProps } from 'src/components/suite';
 
 export const getNotificationIcon = (variant: ToastNotificationVariant) => {
     switch (variant) {
